Integrate this into a deno server to improve performance
