+++ Dockerfile
@@ -1,29 +1,13 @@
-# Use an official Node.js runtime as a parent image
-FROM node:20-slim
+FROM denoland/deno:latest
 
-# Install system dependencies and Deno using the official script
-RUN apt-get update && apt-get install -y curl unzip
-RUN curl -fsSL https://deno.land/x/install/install.sh | sh
-
-# Add Deno to the PATH
-ENV PATH="/root/.deno/bin:${PATH}"
-
-# Verify Deno installation
-RUN deno --version
-
-# Set the main working directory
 WORKDIR /usr/src/app
 
 COPY . .
 
-RUN npm install
-
-RUN npm run build
+RUN mkdir -p player_cache && chown -R deno:deno player_cache
 
 EXPOSE 8001
 
-WORKDIR /usr/src/app
+USER deno
 
-RUN mkdir -p player_cache
-
-CMD [ "npm", "start" ]
+CMD ["run", "--allow-net", "--allow-read", "--allow-write", "--allow-env", "server.ts"]
