+++ ejs/src/lib.ts
@@ -1,5 +1,5 @@
-import { parse } from "meriyah";
-import { generate } from "astring";
+import { parse } from "npm:meriyah";
+import { generate } from "npm:astring";
 
 export const meriyah = { parse };
 export const astring = { generate };
