--- package.json
@@ -1,23 +1,0 @@
-{
-  "name": "yt-cipher-api",
-  "version": "1.0.0",
-  "description": "API for decrypting YouTube signatures using ejs",
-  "main": "dist/api.js",
-  "type": "module",
-  "scripts": {
-    "start": "node dist/api.js",
-    "build": "tsc",
-    "dev": "tsc -w & nodemon dist/api.js"
-  },
-  "dependencies": {
-    "express": "^4.18.2",
-    "node-fetch": "^3.3.2"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.21",
-    "@types/node": "^20.9.0",
-    "@types/tmp": "^0.2.6",
-    "nodemon": "^3.0.1",
-    "typescript": "^5.2.2"
-  }
-}
